# docs/personasSwagger.yaml
openapi: 3.0.3
info:
  title: AKI! - Microservice A (Personas)
  version: "1.0.0"
  description: |
    Personas microservice for AKI! — manages Students, Teachers and Classes.
    Architecture notes: designed following Clean Architecture, SOLID and Vertical Slice.
    This service is intended to be consumed by API Gateway, BFF and other internal services.
  contact:
    name: AKI! Team
    email: infra@aki.example
servers:
  - url: http://localhost:3000
    description: Local development (no /v1 prefix yet)
  - url: https://personas.api.aki.example/v1
    description: Production server
security:
  - bearerAuth: []
...
paths:
...
  /teachers:
    get:
      tags: [Teachers]
      summary: List teachers (paged)
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: size
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
      responses:
        "200":
          description: Paged teachers
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedTeachers'
      security:
        - bearerAuth: []
    post:
      tags: [Teachers]
      summary: Create a new teacher
      description: If created without a `password_hash`, the service emits an internal event `teacher.password.setup` to trigger email setup (async).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TeacherCreate'
      responses:
        "201":
          description: Teacher created (and event emitted if no password)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Teacher'
        "409":
          description: CPF or email conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
        - bearerAuth: []
  /teachers/{teacherId}:
    parameters:
      - name: teacherId
        in: path
        required: true
        schema:
          type: integer
    get:
      tags: [Teachers]
      summary: Get teacher by id
      responses:
        "200":
          description: Teacher found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Teacher'
        "404":
          $ref: '#/components/responses/NotFound'
      security:
        - bearerAuth: []
    put:
      tags: [Teachers]
      summary: Update teacher
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TeacherUpdate'
      responses:
        "200":
          description: Teacher updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Teacher'
        "404":
          $ref: '#/components/responses/NotFound'
      security:
        - bearerAuth: []
    delete:
      tags: [Teachers]
      summary: Delete teacher
      responses:
        "204":
          description: Teacher deleted
        "404":
          $ref: '#/components/responses/NotFound'
      security:
        - bearerAuth: []
...
  /classes/{classId}/teachers/{teacherId}:
    delete:
      tags: [Classes]
      summary: Remove teacher from class
      description: Removes the teacher with the given teacherId from the class with classId.
      parameters:
        - name: classId
          in: path
          required: true
          schema:
            type: integer
        - name: teacherId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Teacher removed
        "404":
          $ref: '#/components/responses/NotFound'
      security:
        - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
    Error:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: array
          items:
            type: string
      example:
        code: "validation_error"
        message: "CPF is required"
        details: ["cpf must be 11 digits"]
...
    Teacher:
      type: object
      properties:
        id:
          type: integer
        cpf:
          type: string
        full_name:
          type: string
        email:
          type: string
        password_hash:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required: [ id, cpf, full_name, email ]
    TeacherCreate:
      type: object
      required: [ cpf, full_name, email ]
      properties:
        cpf:
          type: string
        full_name:
          type: string
        email:
          type: string
        password_hash:
          type: string
          description: Optional - if absent, a `teacher.password.setup` event will be emitted for email setup
    TeacherUpdate:
      type: object
      properties:
        full_name:
          type: string
        email:
          type: string
        password_hash:
          type: string
...
    PagedMeta:
      type: object
      properties:
        page:
          type: integer
        size:
          type: integer
        total:
          type: integer
...
    PagedTeachers:
      type: object
      properties:
        meta:
          $ref: '#/components/schemas/PagedMeta'
        items:
          type: array
          items:
            $ref: '#/components/schemas/Teacher'
...